Change main menu screen and removed creep getting damage sound
change the main menu screen to only have play and quit, which leads to
wave mode. also made them a bit larger. the sound of creeps getting
damage was there again for some reason(had been previously removed cause
it was too annoying) so i removed it again, so now there's only a sound
when the creeps die.